Changed file structure. Tree commands for music now all located in client.py and import from muzique.py
